Обработаны замечания
