+++ src/main/models/Task.java
@@ -99,7 +99,7 @@ public class Task {
 
     @Override
     public int hashCode() {
-        return Objects.hash(title, description, id, status, duration, startTime);
+        return Objects.hash(id);
     }
 
     @Override
