+++ src/main/service/TaskManager.java
@@ -7,7 +7,6 @@ import main.models.Task;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.TreeSet;
 
 public interface TaskManager {
 
@@ -51,5 +50,5 @@ public interface TaskManager {
 
     List<Task> getHistory();
 
-    TreeSet<Task> getPrioritizedTasks();
+    List<Task> getPrioritizedTasks();
 }
